+++ code/.storybook/main.ts
@@ -1,9 +1,11 @@
 import { join } from 'node:path';
 
+import react from '@vitejs/plugin-react';
+
 import { defineMain } from '../frameworks/react-vite/src/node';
 
 const componentsPath = join(__dirname, '../core/src/components/index.ts');
-const managerApiPath = join(__dirname, '../core/src/manager-api/index.ts');
+const managerApiPath = join(__dirname, '../core/src/manager-api/index.mock.ts');
 const imageContextPath = join(__dirname, '../frameworks/nextjs/src/image-context.ts');
 
 const config = defineMain({
@@ -34,20 +36,12 @@ const config = defineMain({
       titlePrefix: 'blocks',
     },
     {
+      directory: '../addons/a11y/src',
+      titlePrefix: 'addons/accessibility',
+    },
+    {
       directory: '../addons/a11y/template/stories',
-      titlePrefix: 'addons/a11y',
-    },
-    {
-      directory: '../addons/backgrounds/template/stories',
-      titlePrefix: 'addons/backgrounds',
-    },
-    {
-      directory: '../addons/controls/src',
-      titlePrefix: 'addons/controls',
-    },
-    {
-      directory: '../addons/controls/template/stories',
-      titlePrefix: 'addons/controls',
+      titlePrefix: 'addons/accessibility',
     },
     {
       directory: '../addons/docs/template/stories',
@@ -58,28 +52,12 @@ const config = defineMain({
       titlePrefix: 'addons/links',
     },
     {
-      directory: '../addons/viewport/template/stories',
-      titlePrefix: 'addons/viewport',
-    },
-    {
-      directory: '../addons/toolbars/template/stories',
-      titlePrefix: 'addons/toolbars',
-    },
-    {
       directory: '../addons/themes/template/stories',
       titlePrefix: 'addons/themes',
     },
     {
       directory: '../addons/onboarding/src',
       titlePrefix: 'addons/onboarding',
-    },
-    {
-      directory: '../addons/interactions/src',
-      titlePrefix: 'addons/interactions',
-    },
-    {
-      directory: '../addons/interactions/template/stories',
-      titlePrefix: 'addons/interactions/tests',
     },
     {
       directory: '../addons/test/src/components',
@@ -92,7 +70,7 @@ const config = defineMain({
   ],
   addons: [
     '@storybook/addon-themes',
-    '@storybook/addon-essentials',
+    '@storybook/addon-docs',
     '@storybook/addon-storysource',
     '@storybook/addon-designs',
     '@storybook/addon-test',
@@ -101,7 +79,6 @@ const config = defineMain({
   ],
   previewAnnotations: [
     './core/template/stories/preview.ts',
-    './addons/toolbars/template/stories/preview.ts',
     './renderers/react/template/components/index.js',
   ],
   build: {
@@ -127,8 +104,6 @@ const config = defineMain({
     disableTelemetry: true,
   },
   features: {
-    viewportStoryGlobals: true,
-    backgroundsStoryGlobals: true,
     developmentModeForBuild: true,
   },
   viteFinal: async (viteConfig, { configType }) => {
@@ -143,9 +118,12 @@ const config = defineMain({
                 'storybook/manager-api': managerApiPath,
                 'sb-original/image-context': imageContextPath,
               }
-            : {}),
+            : {
+                'storybook/manager-api': managerApiPath,
+              }),
         },
       },
+      plugins: [react()],
       optimizeDeps: {
         force: true,
         include: ['@storybook/blocks'],