+++ code/addons/test/src/manager.tsx
@@ -1,35 +1,20 @@
 import React, { useState } from 'react';
 
-import { AddonPanel } from 'storybook/internal/components';
-import type { StatusValue } from 'storybook/internal/types';
 import { type Addon_TestProviderType, Addon_TypesEnum } from 'storybook/internal/types';
 
-import { a11yStatusStore, componentTestStatusStore, store } from '#manager-store';
-import type { Combo } from 'storybook/manager-api';
-import { Consumer, addons, types } from 'storybook/manager-api';
+import {
+  a11yStatusStore,
+  componentTestStatusStore,
+  store,
+  testProviderStore,
+} from '#manager-store';
+import { addons } from 'storybook/manager-api';
 
 import { GlobalErrorContext, GlobalErrorModal } from './components/GlobalErrorModal';
-import { Panel } from './components/Panel';
-import { PanelTitle } from './components/PanelTitle';
+import { SidebarContextMenu } from './components/SidebarContextMenu';
 import { TestProviderRender } from './components/TestProviderRender';
-import {
-  A11Y_PANEL_ID,
-  ADDON_ID,
-  type Details,
-  PANEL_ID,
-  STATUS_TYPE_ID_A11Y,
-  STATUS_TYPE_ID_COMPONENT_TEST,
-  TEST_PROVIDER_ID,
-} from './constants';
-import type { TestStatus } from './node/reporter';
-
-const statusMap: Record<TestStatus, StatusValue> = {
-  pending: 'status-value:pending',
-  passed: 'status-value:success',
-  warning: 'status-value:warning',
-  failed: 'status-value:error',
-  skipped: 'status-value:unknown',
-};
+import { A11Y_PANEL_ID, ADDON_ID, PANEL_ID, TEST_PROVIDER_ID } from './constants';
+import { useTestProvider } from './use-test-provider-state';
 
 addons.register(ADDON_ID, (api) => {
   const storybookBuilder = (globalThis as any).STORYBOOK_BUILDER || '';
@@ -44,125 +29,69 @@ addons.register(ADDON_ID, (api) => {
     a11yStatusStore.onSelect(() => {
       openPanel(A11Y_PANEL_ID);
     });
+    testProviderStore.onRunAll(() => {
+      store.send({
+        type: 'TRIGGER_RUN',
+        payload: {
+          triggeredBy: 'run-all',
+        },
+      });
+    });
+    store.untilReady().then(() => {
+      store.setState((state) => ({
+        ...state,
+        indexUrl: new URL('index.json', window.location.href).toString(),
+      }));
+    });
 
     addons.add(TEST_PROVIDER_ID, {
       type: Addon_TypesEnum.experimental_TEST_PROVIDER,
-      runnable: true,
-      name: 'Component tests',
-      // @ts-expect-error: TODO: Fix types
-      render: (state) => {
+      render: () => {
         const [isModalOpen, setModalOpen] = useState(false);
+        const {
+          storeState,
+          setStoreState,
+          testProviderState,
+          componentTestStatusValueToStoryIds,
+          a11yStatusValueToStoryIds,
+          isSettingsUpdated,
+        } = useTestProvider(api);
         return (
-          <GlobalErrorContext.Provider
-            value={{ error: state.error?.message, isModalOpen, setModalOpen }}
-          >
-            <TestProviderRender api={api} state={state} />
+          <GlobalErrorContext.Provider value={{ isModalOpen, setModalOpen }}>
+            <TestProviderRender
+              api={api}
+              storeState={storeState}
+              setStoreState={setStoreState}
+              isSettingsUpdated={isSettingsUpdated}
+              testProviderState={testProviderState}
+              componentTestStatusValueToStoryIds={componentTestStatusValueToStoryIds}
+              a11yStatusValueToStoryIds={a11yStatusValueToStoryIds}
+            />
             <GlobalErrorModal
+              storeState={storeState}
               onRerun={() => {
                 setModalOpen(false);
-                api.runTestProvider(TEST_PROVIDER_ID);
+                store.send({
+                  type: 'TRIGGER_RUN',
+                  payload: {
+                    triggeredBy: 'global',
+                  },
+                });
               }}
             />
           </GlobalErrorContext.Provider>
         );
       },
 
-      // @ts-expect-error: TODO: Fix types
-      sidebarContextMenu: ({ context, state }) => {
+      sidebarContextMenu: ({ context }) => {
         if (context.type === 'docs') {
           return null;
         }
         if (context.type === 'story' && !context.tags.includes('test')) {
           return null;
         }
-        return (
-          <TestProviderRender
-            api={api}
-            state={state}
-            entryId={context.id}
-            style={{ minWidth: 240 }}
-          />
-        );
+        return <SidebarContextMenu context={context} api={api} />;
       },
-
-      // @ts-expect-error: TODO: Fix types
-      stateUpdater: (state, update) => {
-        const updated = {
-          ...state,
-          ...update,
-          details: { ...state.details, ...update.details },
-        };
-
-        if ((!state.running && update.running) || store.getState().watching) {
-          // Clear coverage data when starting test run or enabling watch mode
-          delete updated.details.coverageSummary;
-        }
-
-        if (update.details?.testResults) {
-          componentTestStatusStore.set(
-            update.details.testResults.flatMap((testResult) =>
-              testResult.results
-                .filter(({ storyId }) => storyId)
-                .map(({ storyId, status, testRunId, ...rest }) => {
-                  return {
-                    storyId,
-                    typeId: STATUS_TYPE_ID_COMPONENT_TEST,
-                    value: statusMap[status],
-                    title: 'Component tests',
-                    description:
-                      'failureMessages' in rest && rest.failureMessages
-                        ? rest.failureMessages.join('\n')
-                        : '',
-                    data: { testRunId },
-                    sidebarContextMenu: false,
-                  };
-                })
-            )
-          );
-          a11yStatusStore.set(
-            update.details.testResults.flatMap((testResult) =>
-              testResult.results
-                .filter(({ storyId, reports }) => {
-                  const a11yReport = reports.find((r: any) => r.type === 'a11y');
-                  return storyId && a11yReport;
-                })
-                .map(({ storyId, testRunId, reports }) => {
-                  const a11yReport = reports.find((r: any) => r.type === 'a11y')!;
-                  return {
-                    storyId,
-                    typeId: STATUS_TYPE_ID_A11Y,
-                    value: statusMap[a11yReport.status],
-                    title: 'Accessibility tests',
-                    description: '',
-                    data: { testRunId },
-                    sidebarContextMenu: false,
-                  };
-                })
-            )
-          );
-        }
-
-        return updated;
-      },
-    } satisfies Omit<Addon_TestProviderType<Details>, 'id'>);
+    });
   }
-
-  const filter = ({ state }: Combo) => {
-    return {
-      storyId: state.storyId,
-    };
-  };
-
-  addons.add(PANEL_ID, {
-    type: types.PANEL,
-    title: () => <PanelTitle />,
-    match: ({ viewMode }) => viewMode === 'story',
-    render: ({ active }) => {
-      return (
-        <AddonPanel active={!!active}>
-          <Consumer filter={filter}>{({ storyId }) => <Panel storyId={storyId} />}</Consumer>
-        </AddonPanel>
-      );
-    },
-  });
 });