+++ code/addons/viewport/src/index.ts
@@ -1,8 +1,0 @@
-import { definePreview } from 'storybook/preview-api';
-
-import * as addonAnnotations from './preview';
-
-export * from './defaults';
-export type * from './types';
-
-export default () => definePreview(addonAnnotations);