+++ code/core/src/common/presets.test.ts
@@ -38,7 +38,6 @@ vi.mock('./utils/safeResolve', () => {
     '@storybook/addon-docs',
     '@storybook/addon-cool',
     '@storybook/addon-docs/preset',
-    '@storybook/addon-interactions/preset',
     '@storybook/addon-essentials',
     '@storybook/addon-knobs/manager',
     '@storybook/addon-knobs/register',
@@ -446,10 +445,9 @@ describe('loadPreset', () => {
     mockPreset('@storybook/addon-docs/preset', {});
     mockPreset('addon-foo/register.js', {});
     mockPreset('@storybook/addon-cool', {});
-    mockPreset('@storybook/addon-interactions/preset', {});
     mockPreset('addon-bar', {
       addons: ['@storybook/addon-cool'],
-      presets: ['@storybook/addon-interactions/preset'],
+      presets: [],
     });
     mockPreset('addon-baz/register.js', {});
     mockPreset('@storybook/addon-notes/register-panel', {});
@@ -540,11 +538,6 @@ describe('loadPreset', () => {
         preset: {
           managerEntries: [normalize('addon-foo/register')],
         },
-      },
-      {
-        name: '@storybook/addon-interactions/preset',
-        options: {},
-        preset: {},
       },
       {
         name: '@storybook/addon-cool',
@@ -664,11 +657,6 @@ describe('loadPreset', () => {
       [
         {
           "name": "@storybook/preset-typescript",
-          "options": {},
-          "preset": {},
-        },
-        {
-          "name": "@storybook/addon-interactions/preset",
           "options": {},
           "preset": {},
         },