+++ code/core/src/instrumenter/index.ts
@@ -1,3 +1,4 @@
-export { EVENTS, instrument } from './instrumenter';
+export { instrument } from './instrumenter';
 
+export { EVENTS } from './EVENTS';
 export * from './types';