+++ code/core/src/manager/components/sidebar/Heading.tsx
@@ -2,7 +2,6 @@ import type { ComponentProps, FC } from 'react';
 import React from 'react';
 
 import { Button } from 'storybook/internal/components';
-import type { Addon_SidebarTopType } from 'storybook/internal/types';
 
 import { styled } from 'storybook/theming';
 
@@ -13,7 +12,6 @@ import { SidebarMenu } from './Menu';
 export interface HeadingProps {
   menuHighlighted?: boolean;
   menu: MenuList;
-  extra: Addon_SidebarTopType[];
   skipLinkHref?: string;
   isLoading: boolean;
   onMenuClick?: SidebarMenuProps['onClick'];
@@ -83,7 +81,6 @@ export const Heading: FC<HeadingProps & ComponentProps<typeof HeadingWrapper>> =
   menuHighlighted = false,
   menu,
   skipLinkHref,
-  extra,
   isLoading,
   onMenuClick,
   ...props
@@ -102,7 +99,6 @@ export const Heading: FC<HeadingProps & ComponentProps<typeof HeadingWrapper>> =
         <Brand />
       </BrandArea>
 
-      {isLoading ? null : extra.map(({ id, render: Render }) => <Render key={id} />)}
       <SidebarMenu menu={menu} isHighlighted={menuHighlighted} onClick={onMenuClick} />
     </HeadingWrapper>
   );