+++ test-storybooks/server-kitchen-sink/.storybook/main.ts
@@ -6,10 +6,7 @@ const mainConfig: StorybookConfig = {
   addons: [
     '@storybook/addon-docs',
     '@storybook/addon-a11y',
-    '@storybook/addon-backgrounds',
-    '@storybook/addon-links',
-    '@storybook/addon-controls',
-    '@storybook/addon-highlight',
+    '@storybook/addon-links'
   ],
   core: {
     disableTelemetry: true,